Add files via upload
stx_det_cal_proc  is the entry point into the demonstration of simulating a baseline and offset calibration spectrum. It also processes the spectra through the compression and decompression and finally determines the spectrum offset in keV.  The energy offset is used to determine the elut 4096 channel offset needed for setting the science energy channels. The spectra are constructed using an Monte Carlo procedure (or not) within stx_bkg_sim_spectrum.  The process of telemetry compression and decompression is done within stx_cal_spec_telem based on the grouping configuration in stx_bkg_ecal_spec_config.  Fitting through Ospex is managed in stx_det_cal_proc with added support from stx_cal_script.  There is also a cross-correlation module that's been added in the util branch.