These files track the energy/channel status of the 384 individual detectors
Stx_science_energy_channels gives the energy edge targets for the ADC accumulation bins. That, with the results of the calibration spectra, specifies the ELUT which gets loaded by a separate process into telecommands to control the action of the FPGA on the instrument in space on Solar Orbiter. The fitting process is controlled by  stx_calib_fit.pro A new ELUT can be written in the working directory at the discretion of the caller. Or using the results file from that process the new ELUT can be written.