Add files via upload
stx_calib_groups.pro and stix_calib_groups.ps document the initial selection of the detectors and channel ranges to use at full and partial resolution  for the calibration  downloads.  There are 3 groups of detectors, the center group (based on adc1024 channel values) contains most of the detectors.  Those groups are for the 31 and 81 keV lines which are fitted in the calibration task for the Ba 133 cal source lines. In addition , for each group there is a center and two wing groups with 4 channel grouped binning. After the initial downloads, and occasional downloads at full resolution these are the 5 subgroups to be defined for download.