Add files via upload
This archive contains a survey of the 2019 Sept Munich calibration runs. 21 runs in all.  There are plots in the calibration range, 300-550 ADC 1024,  for all 384 detectors for all 21 runs. All the results are contained in a csv file for each run.  Finally, there are a selected 75 high quality plots and a csv report on the same showing the nominal deviation in the 30.85 keV calibration line in units of ADC1024.  These 75 spectra are of high quality and I don't know why all of the detectors weren't producing this quality spectrum.