Add files via upload
This code is used to build a database file of the processed calibration FITS files (L1) and from this compute the derivative of the ADC4096 channel vs temperature change. This derivative is exactly what is used to build the TLUT but the current temperature range and channel variability is too small to warrant a TLUT at this time. 6-aug-202