Update README.md - Who to contact
I was thinking if in the table "who to contact" we can add the imaging-spectroscopy topic. It is something I am working on and planning to have a first version by autumn. In this way, we can show we are working on it and if someone is interested in it can contact me. It is a tricky task, so it is better if other users first contact us before publishing something on imaging-spectroscopy.