Update stx_estimate_flare_location
This new version provides an estimate of the reliability of the estimated flare location as the ratio between the main peak of the full-disk BP and the peak of the sidelobes. In the case when this ratio exceeds a specific threshold (0.9), the procedure prints a waning since the estimated location could be unreliable.